requirements.txt
fastapi
uvicorn
newspaper3k
pydantic
requests
lxml_html_clean
openai
python-dotenv
pytest